Create SecondsAndMinutes.java
The purpose of this Java program is to take some input in the form of minutes or seconds, and then convert it to the corresponding hours + minutes + seconds. The second method of the program takes only seconds, and then it converts it to hours and minutes and seconds. 
CHEERS. HAPPY CODING WITH JAVA.