Create NumberOfDaysInOneMonth
This Java program demonstrates the use of switch and if-else statements to show if a year is leap or not. If a year is leap, then it will return the number of days for a specific month. Otherwise, if it's not a leap year then it will return a value of -1, indicating an invalid value.