Create ForStatement_Count_Sum_Numbers
This Java program is an example in the use of for loops. We are trying to sum the first five numbers in the range [1, 1000], that are divisible by both 3 and 5. After we get all those numbers we want to store  them in a variable and then later we can print out to the console the sum of all those numbers.
#HappyCoding