Create SumOfEvenOddNumbers
This Java program checks to see if the numbers entered by the user are even or odd. If they are even then display in the console the sum of all even numbers which were entered by the user. Otherwise, if the numbers are odd, then it will display to the console the odd numbers. Actually, the progamm will display both even and odd numbers.
Happy coding