Create ComplexNumber
This Java program will add and subtract two complex numbers. A complex number is any number of the form a + b*i, where a, b are real numbers, and i is called the imaginary unit (equivalent to √(-1). The a is known as the real part, whereas the b is known as the imaginary part (it's being multiplied by the imaginary unit, hence an imaginary part).
#HappyCoding