Create Printer
This Java program is a demonstration of a printer in real life. It allows you to print paper if and only if the level of the toner allows it. It will also keep count of the pages printed by the printer. But it doesn't take in count if the printer is duplex or not. So this printer will only count sheets of papers printed and not take into consideration the fact that each page has two sheets. Anyway if you want to be more specific with it, you can copy the code and experiment with it in any way you want.
#Happy Coding